# uint256: optimize leadingZeros and add BenchmarkBitLen