--- a/mod.go
+++ b/mod.go
@@ -10,13 +10,8 @@ import (
 
 // Some utility functions
 
-func leadingZeros(x *Int) (z int) {
-	var t int
-	z = bits.LeadingZeros64(x[3])
-	t = bits.LeadingZeros64(x[2]); if z ==  64 { z = t +  64 }
-	t = bits.LeadingZeros64(x[1]); if z == 128 { z = t + 128 }
-	t = bits.LeadingZeros64(x[0]); if z == 192 { z = t + 192 }
-	return z
+func leadingZeros(x *Int) int {
+	return 256 - x.BitLen()
 }
 
 // Reciprocal computes a 320-bit value representing 1/m